Refactor getPost function to use doFetch helper in getPost.mjs
