+++ js/listeners/updatePost.mjs
@@ -1,4 +1,4 @@
-import { getPost } from "../api/posts/read.mjs";
+import { getPost } from "../api/posts/getPost.mjs";
 import { updatePost } from "../api/posts/update.mjs";
